refactor: create pages folder and move the Main component there
